refactor(admin): migrate ?tab= query param to nested routes
- Add app/admin/layout.tsx with SubNav using <Link> components
- Split admin sections into separate page files:
  /admin           → reservations (default)
  /admin/hygiene   → odometer gap hygiene
  /admin/wagens    → fleet / cars
  /admin/leden     → members
  /admin/afrek     → settlement
  /admin/payout    → owner payout
- Extract shared types/hooks to app/admin/_shared.tsx
- Remove useQueryParam("tab") and ?tab= logic

Closes #13
